Version 0.8c build 000008c - Update + dpg support
New Features:
- DPG (Dyanmic Pages) allows Pages module to create a pages xml along
with an extended php page just for that xml. If a page is dpg then the
php file content will override the xml. This allows you to create a new
page and use code for its content rather than using the ckeditor for
page's pages. This feature will allow other modules to create custom
pages that work within pageLoader and menuLoader in the site template.
- A new Pages editor just for dpg files.