Version 0.9c build 000009c - Prepairing for version 1.0 Alpha stable release
What's new?
1. Switches, super's version of a registry.
2. Welcome page shows up if it is your first time powering up super.
Also allows you to configure super to get started in just a few clicks.
3. Online feature for Security (formerly People) shows you who is
currently online on your super installation. Dont confuse this with
users, It will show the computer names of the computers that have logged
on. Not the username. Two sessions on one computer will show up as just
one.
4. View site option now in the Admin MenuDropDown menu.
5. Security/users and Security/groups are now backed up in the backup
tool in the Settings panel.
6. Brute force logout is now a sitewide feature working with group
permissions.
7. Super version number and build number are now loaded from a switch
called "version". for easier access.